api/requirements.txt
fastapi
uvicorn
python-multipart
pillow
pix2tex
torch
torchvision
timm
requests
numpy
pandas
sympy
jinja2
pydantic
safetensors
fsspec
h11
click
certifi
anyio
idna
regex
tqdm
transformers
einops
entmax
scipy
opencv-python
albucore
munch
pytz
tzdata
markupsafe
filelock
six
